+++ test/index.test.js
@@ -1,0 +1,4 @@
+require('./sync')
+require('./alterColumns')
+require('./reconnection')
+require('./schemaValidation')